backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.39.0
sqlalchemy==2.0.48
pydantic==2.12.5
python-dotenv==1.2.1
psycopg2-binary==2.9.11
redis==7.0.1
python-jose==3.5.0
passlib==1.7.4